// apps/bsky_bot_server/src/types.ts
type ProfileView = AppBskyActorDefs.ProfileView;
type ProfileViewDetailed = AppBskyActorDefs.ProfileViewDetailed;
// ...
export type LanguageName =
  | "Arabic"
  | "Bengali"
  | "Bulgarian"
  | "Chinese"
  | "Croatian"
  | "Czech"
  | "Danish"
  | "Dutch"
  | "English"
  | "Estonian"
  | "Finnish"
  | "French"
  | "German"
  | "Greek"
  | "Hebrew"
  | "Hindi"
  | "Hungarian"
  | "Indonesian"
  | "Italian"
  | "日本語"
  | "Korean"
  | "Latvian"
  | "Lithuanian"
  | "Norwegian"
  | "Polish"
  | "Portuguese"
  | "Romanian"
  | "Russian"
  | "Serbian"
  | "Slovak"
  | "Slovenian"
  | "Spanish"
  | "Swahili"
  | "Swedish"
  | "Thai"
  | "Turkish"
  | "Ukrainian"
  | "Vietnamese";
// ...
export type UserInfoGemini = {
  follower: ProfileView;
  langStr?: LanguageName;
  posts?: string[];
  lastYearPosts?: string[];
  likedByFollower?: string[];
  history?: Content[];
  image?: ImageRef[];
  anniversary?: Holiday[];
  followersFriend?: {
    profile: ProfileView;
    post?: string;
  }[];
  embed?: Embed;
  topWords?: string[];
  postOnMonth?: string[][];
};

export interface Embed {
  profile_embed?: ProfileViewDetailed;
  text_embed?: string;
  uri_embed?: string;
  image_embed?: ImageRef[];
}

export interface ImageRef {
  image_url: string;
  mimeType: string;
}

export interface FeatureContext {
  [key: string]: any;
}

export type GeminiResponseResult = string | {
  text: string;
  imageBlob?: BlobRef;
  embedTo?: {
    uri: string;
    cid: string;
  }
};

export type GeminiScore = {
  comment: string;
  score: number;
}

export interface ThreadInfo {
  parent?: PostRecord
  grandParent?: PostRecord
  root?: PostRecord
}

export type Holiday = {
  id: string;
  names: { ja: string; en: string };
  rule:
  | { type: "fixed"; month: number; day: number }
  | { type: "nth-weekday"; month: number; week: number; weekday: number } // weekday: 0=Sun..6=Sat
  | { type: "easter"; calendar: "western" | "orthodox" };
  regions?: string[];
}